requirements.txt
beautifulsoup4==4.9.3
PyNaCl
urllib3
ssl
Wavelink
DiscordUtils==1.3.4
discord_components==1.1.4
asyncpg
matplotlib==3.2.0
numpy
aiohttp==3.6.3
valoStatus==3.2.15
